chore(refactor): AccordianDetails to AccordionDetails
