+++ src/components/ui/TripCard/AccordionDetails/Checkout/Checkout.styled.ts
