+++ src/components/ui/TripCard/AccordionDetails/Checkout/Checkout.tsx
