+++ src/components/ui/TripCard/AccordionDetails/DetailsGrid/DetailItem/DetailItem.tsx
