+++ src/components/ui/TripCard/AccordionDetails/DetailsGrid/DetailsGrid.tsx
