+++ src/components/ui/TripCard/AccordionDetails/SeatLegend/SeatLegend.styled.ts
