+++ src/components/ui/TripCard/AccordionDetails/SeatLegend/SeatLegend.tsx
