+++ src/components/ui/TripCard/AccordionDetails/TripCardDetails.tsx
